Added support for the TCK
This adds support for running against the TCK. See

https://github.com/eclipse/microprofile-reactive/pull/52

for details.

I also switched from gradle to maven. Gradle was just causing too many
problems, it's really not that good as a library developer. Also,
integration with IntelliJ was a problem because it was putting resources
in a different directory to compiled classes, which meant that when the
Weld container started up, it would scan the resources not the classes
directory for beans, and find none.